refactor: Move jobs table into JobsTableView class
